requirements.txt
accelerate==0.34.2
addict==2.4.0
aiohttp==3.10.8
albumentations==1.4.16
anime_face_detector==0.0.9
basicsr==1.4.2
bitsandbytes==0.44.1
caffe2==0.8.1
cityscapesscripts==2.2.4
controlnet_aux==0.0.9
cv_bridge==1.13.0.post0
dadaptation==3.2
diffusers==0.30.3
einops==0.8.0
external==0.0.1
fairscale==0.4.13
ftfy==6.2.3
fvcore==0.1.5.post20221221
gradio==4.44.1
h5py==3.11.0
huggingface_hub==0.25.1
hydra-core==1.3.2
imutils==0.5.4
iopath==0.1.10
k_diffusion==0.1.1.post1
lion_pytorch==0.2.2
llava==0.0.1.dev0
lmdb==1.5.1
lpips==0.1.4
lycoris==0.0.0
matplotlib==3.9.2
mc==0.0
mmdet==3.3.0
numpy==2.1.1
ollama==0.3.3
omegaconf==2.3.0
onnx==1.16.2
onnxruntime==1.19.2
openvino==2024.4.0
optimum==1.22.0
oyaml==1.0
packaging==24.1
pandas==2.2.3
parrots==1.0.3
Pillow==10.4.0
prettytable==3.11.0
prodigyopt==1.0
pycocotools==2.0.8
pydantic==2.9.2
Pygments==2.18.0
python-dotenv==1.0.1
python-slugify==8.0.4
pytorch_fid==0.3.0
pytorch_lightning==2.4.0
PyYAML==6.0.2
PyYAML==6.0.2
regex==2024.9.11
Requests==2.32.3
safetensors==0.4.5
scipy==1.14.1
setuptools==72.1.0
Shapely==2.0.6
skimage==0.0
tabulate==0.9.0
tensorboardX==2.6.2.2
termcolor==2.4.0
tifffile==2024.9.20
timm==1.0.9
toml==0.10.2
torch==2.4.1
torch_tb_profiler==0.4.3
torch_xla==2.4.0
torchvision==0.19.1
tqdm==4.66.4
transformers==4.45.1
turbojpeg==0.0.2
typing_extensions==4.12.2
voluptuous==0.15.2
xformers==0.0.28.post1
yapf==0.40.2
